Extract Youtube Responses to own trait
