+++ tests/Feature/Jobs/ImportYoutubeChannelStreamsJobTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Feature\Jobs;
 
+use Tests\Fakes\YoutubeReponses;
 use App\Jobs\ImportYoutubeChannelStreamsJob;
 use App\Models\Channel;
 use App\Models\Stream;
@@ -12,6 +13,7 @@ use Tests\TestCase;
 class ImportYoutubeChannelStreamsJobTest extends TestCase
 {
     use RefreshDatabase;
+    use YoutubeReponses;
 
     /** @test */
     public function it_imports_youtube_channel_streams(): void
