+++ tests/Unit/YoutubeTest.php
@@ -3,12 +3,15 @@
 namespace Tests\Unit;
 
 use App\Facades\Youtube;
+use Tests\Fakes\YoutubeReponses;
 use App\Services\Youtube\StreamData;
 use Illuminate\Support\Facades\Http;
 use Tests\TestCase;
 
 class YoutubeTest extends TestCase
 {
+    use YoutubeReponses;
+
     /** @test */
     public function it_can_fetch_channel_details_from_youtube(): void
     {
